src/main/java/guru/springframework/sfgpetclinic/services/springdatajpa/VisitSDJpaService.java
package guru.springframework.sfgpetclinic.services.springdatajpa;

import guru.springframework.sfgpetclinic.model.Visit;
import guru.springframework.sfgpetclinic.repositories.VisitRepository;
import guru.springframework.sfgpetclinic.services.VisitService;

import java.util.HashSet;
import java.util.Set;

public class VisitSDJpaService implements VisitService {

    private final VisitRepository visitRepository;

    public VisitSDJpaService(VisitRepository visitRepository) {
        this.visitRepository = visitRepository;
    }

    @Override
    public Set<Visit> findAll33() {
        Set<Visit> visits = new HashSet<>();
        visitRepository.findAll().forEach(visits::add);
        return visits;
    }

    @Override
    public Visit findById33(Long aLong) {
        return visitRepository.findById(aLong).orElse(null);
    }

    @Override
    public Visit save33(Visit object) {
        return visitRepository.save(object);
    }

    @Override
    public void delete33(Visit object) {
        visitRepository.delete(object);
    }

    @Override
    public void deleteById33(Long aLong) {
        visitRepository.deleteById(aLong);
    }
}
